+++ src/PlotGen.cpp
@@ -6,7 +6,7 @@ namespace StopPow
 
 // Generate a plot of dEdx vs E for a given model. Results stored in data.
 bool get_dEdx_vs_E( 	StopPow & model ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	// call overloaded function using default step size:
 	return get_dEdx_vs_E(	model,
@@ -16,7 +16,7 @@ bool get_dEdx_vs_E( 	StopPow & model ,
 // Generate a plot of dEdx vs E for a given model. Results stored in data.
 bool get_dEdx_vs_E( 	StopPow & model ,
 						int num_points ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	// call overloaded function using model's energy limits:
 	return get_dEdx_vs_E(	model,
@@ -27,24 +27,24 @@ bool get_dEdx_vs_E( 	StopPow & model ,
 }
 // Generate a plot of dEdx vs E for a given model. Results stored in data.
 bool get_dEdx_vs_E( 	StopPow & model ,
-						float Emin ,
-						float Emax ,
+						double Emin ,
+						double Emax ,
 						int num_points ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> dEdx_vector;
+	std::vector<double> E_vector; std::vector<double> dEdx_vector;
 	data.push_back( E_vector );
 	data.push_back( dEdx_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float E = Emin; E <= Emax; E+=dE )
+	for(double E = Emin; E <= Emax; E+=dE )
 	{
 		try
 		{
@@ -63,7 +63,7 @@ bool get_dEdx_vs_E( 	StopPow & model ,
 
 // generate a plot of range vs E for a given model. Results stored in data
 bool get_Range_vs_E( 	StopPow & model ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	return get_Range_vs_E(	model ,
 							PLOT_DEFAULT_NUM_POINTS ,
@@ -73,7 +73,7 @@ bool get_Range_vs_E( 	StopPow & model ,
 // generate a plot of range vs E for a given model. Results stored in data
 bool get_Range_vs_E( 	StopPow & model ,
 						int num_points ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	return get_Range_vs_E(	model,
 							model.get_Emin(),
@@ -84,24 +84,24 @@ bool get_Range_vs_E( 	StopPow & model ,
 
 // generate a plot of range vs E for a given model. Results stored in data
 bool get_Range_vs_E( 	StopPow & model ,
-						float Emin ,
-						float Emax ,
+						double Emin ,
+						double Emax ,
 						int num_points ,
-						std::vector< std::vector<float> > & data )
+						std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Range_vector;
+	std::vector<double> E_vector; std::vector<double> Range_vector;
 	data.push_back( E_vector );
 	data.push_back( Range_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float E = Emin; E <= Emax; E+=dE )
+	for(double E = Emin; E <= Emax; E+=dE )
 	{
 		try
 		{
@@ -113,12 +113,12 @@ bool get_Range_vs_E( 	StopPow & model ,
 			}
 			else
 			{
-				float Range_last = data[1].back();
+				double Range_last = data[1].back();
 
 				data[0].push_back(E);
 				// instead of calculating whole range, use previous result
 				// to speed up calculation
-				float range = Range_last + model.Thickness(E, fmax(E-dE,Emin) );
+				double range = Range_last + model.Thickness(E, fmax(E-dE,Emin) );
 				data[1].push_back( range );
 			}
 		}
@@ -134,8 +134,8 @@ bool get_Range_vs_E( 	StopPow & model ,
 
 // Eout vs Ein given thickness
 bool get_Eout_vs_Ein(	StopPow & model ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {	
 	return get_Eout_vs_Ein(	model,
 							PLOT_DEFAULT_NUM_POINTS,
@@ -146,12 +146,12 @@ bool get_Eout_vs_Ein(	StopPow & model ,
 // Eout vs Ein given thickness
 bool get_Eout_vs_Ein(	StopPow & model ,
 						int num_points ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {	
 	// calculate limits. Lower cutoff based on particle
 	// that just makes it through
-	float Emin, Emax;
+	double Emin, Emax;
 	try
 	{
 		Emin = model.Ein(model.get_Emin(),Thickness);
@@ -171,25 +171,25 @@ bool get_Eout_vs_Ein(	StopPow & model ,
 
 // Eout vs Ein given thickness
 bool get_Eout_vs_Ein(	StopPow & model ,
-						float Emin ,
-						float Emax ,
+						double Emin ,
+						double Emax ,
 						int num_points ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {	
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Eout_vector;
+	std::vector<double> E_vector; std::vector<double> Eout_vector;
 	data.push_back( E_vector );
 	data.push_back( Eout_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float Ein = Emin; Ein <= Emax; Ein+=dE )
+	for(double Ein = Emin; Ein <= Emax; Ein+=dE )
 	{
 		try
 		{
@@ -208,8 +208,8 @@ bool get_Eout_vs_Ein(	StopPow & model ,
 
 // Calculate Eout vs thickness for given input energy
 bool get_Eout_vs_Thickness(	StopPow & model ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	return get_Eout_vs_Thickness(	model,
 									PLOT_DEFAULT_NUM_POINTS,
@@ -220,12 +220,12 @@ bool get_Eout_vs_Thickness(	StopPow & model ,
 // Calculate Eout vs thickness for given input energy
 bool get_Eout_vs_Thickness(	StopPow & model ,
 							int num_points ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	// calculate limits. Lower cutoff based on particle
 	// that just makes it through
-	float Tmin, Tmax;
+	double Tmin, Tmax;
 	try
 	{
 		Tmin = 0;
@@ -245,25 +245,25 @@ bool get_Eout_vs_Thickness(	StopPow & model ,
 
 // Calculate Eout vs thickness for given input energy
 bool get_Eout_vs_Thickness(	StopPow & model ,
-							float Tmin ,
-							float Tmax ,
+							double Tmin ,
+							double Tmax ,
 							int num_points ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Thickness_vector;
+	std::vector<double> E_vector; std::vector<double> Thickness_vector;
 	data.push_back( E_vector );
 	data.push_back( Thickness_vector );
 
 	// step size:
-	float dT = (Tmax-Tmin) / ((float)num_points);
+	double dT = (Tmax-Tmin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float T = Tmin; T <= Tmax; T+=dT )
+	for(double T = Tmin; T <= Tmax; T+=dT )
 	{
 		try
 		{
@@ -275,7 +275,7 @@ bool get_Eout_vs_Thickness(	StopPow & model ,
 			}
 			else
 			{
-				float Elast = data[1].back();
+				double Elast = data[1].back();
 				// do an incremental calculation, which is faster
 				data[0].push_back(T);
 				data[1].push_back( model.Eout(Elast,dT) );
@@ -293,8 +293,8 @@ bool get_Eout_vs_Thickness(	StopPow & model ,
 
 // Get dataset for Ein vs Eout
 bool get_Ein_vs_Eout(	StopPow & model ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {
 	return get_Ein_vs_Eout(	model,
 							PLOT_DEFAULT_NUM_POINTS,
@@ -305,13 +305,13 @@ bool get_Ein_vs_Eout(	StopPow & model ,
 // Get dataset for Ein vs Eout
 bool get_Ein_vs_Eout(	StopPow & model ,
 						int num_points ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {
 	// calculate energy limits
 	// lower limit = particle almost ranged out
 	// upper limit = Emax shifted through Thickness
-	float Emin, Emax;
+	double Emin, Emax;
 	try
 	{
 		Emin = model.get_Emin();
@@ -332,25 +332,25 @@ bool get_Ein_vs_Eout(	StopPow & model ,
 
 // Get dataset for Ein vs Eout
 bool get_Ein_vs_Eout(	StopPow & model ,
-						float Emin ,
-						float Emax ,
+						double Emin ,
+						double Emax ,
 						int num_points ,
-						float Thickness ,
-						std::vector< std::vector<float> > & data )
+						double Thickness ,
+						std::vector< std::vector<double> > & data )
 {	
 	// clear data
 	data.clear();
 
 	// set up 2 vectors
-	std::vector<float> E_vector; std::vector<float> Ein_vector;
+	std::vector<double> E_vector; std::vector<double> Ein_vector;
 	data.push_back( E_vector );
 	data.push_back( Ein_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float Eout = Emin; Eout <= Emax; Eout+=dE )
+	for(double Eout = Emin; Eout <= Emax; Eout+=dE )
 	{
 		try
 		{
@@ -370,8 +370,8 @@ bool get_Ein_vs_Eout(	StopPow & model ,
 
 // Get dataset for Ein vs thickness
 bool get_Ein_vs_Thickness(	StopPow & model ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	return get_Ein_vs_Thickness(model,
 								PLOT_DEFAULT_NUM_POINTS,
@@ -382,13 +382,13 @@ bool get_Ein_vs_Thickness(	StopPow & model ,
 // Get dataset for Ein vs thickness
 bool get_Ein_vs_Thickness(	StopPow & model ,
 							int num_points ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	// calculate limits. Lower cutoff is zero
 	// Upper cutoff corresponds to thickness that ranges
 	// the max model energy to the specified Eout
-	float Tmin, Tmax;
+	double Tmin, Tmax;
 	try
 	{
 		Tmin = 0;
@@ -408,25 +408,25 @@ bool get_Ein_vs_Thickness(	StopPow & model ,
 
 // Get dataset for Ein vs thickness
 bool get_Ein_vs_Thickness(	StopPow & model ,
-							float Tmin ,
-							float Tmax ,
+							double Tmin ,
+							double Tmax ,
 							int num_points ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Thickness_vector;
+	std::vector<double> E_vector; std::vector<double> Thickness_vector;
 	data.push_back( E_vector );
 	data.push_back( Thickness_vector );
 
 	// step size:
-	float dT = (Tmax-Tmin) / ((float)num_points);
+	double dT = (Tmax-Tmin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float T = Tmin; T <= Tmax; T+=dT )
+	for(double T = Tmin; T <= Tmax; T+=dT )
 	{
 		try
 		{
@@ -437,8 +437,8 @@ bool get_Ein_vs_Thickness(	StopPow & model ,
 			}
 			else
 			{
-				float Ein_last = data[1].back();
-				float Ein = model.Ein(Ein_last,dT);
+				double Ein_last = data[1].back();
+				double Ein = model.Ein(Ein_last,dT);
 				data[0].push_back(T);
 				data[1].push_back(Ein);
 			}
@@ -455,8 +455,8 @@ bool get_Ein_vs_Thickness(	StopPow & model ,
 
 // Get data for thickess and a function of energy out given energy in
 bool get_Thickness_vs_Eout(	StopPow & model ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	return get_Thickness_vs_Eout(model,
 								PLOT_DEFAULT_NUM_POINTS,
@@ -467,8 +467,8 @@ bool get_Thickness_vs_Eout(	StopPow & model ,
 // Get data for thickess and a function of energy out given energy in
 bool get_Thickness_vs_Eout(	StopPow & model ,
 							int num_points ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	// energy limits
 	// lower limit is 0 (i.e. ranged out)
@@ -483,25 +483,25 @@ bool get_Thickness_vs_Eout(	StopPow & model ,
 
 // Get data for thickess and a function of energy out given energy in
 bool get_Thickness_vs_Eout(	StopPow & model ,
-							float Emin ,
-							float Emax ,
+							double Emin ,
+							double Emax ,
 							int num_points ,
-							float Ein ,
-							std::vector< std::vector<float> > & data )
+							double Ein ,
+							std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Thickness_vector;
+	std::vector<double> E_vector; std::vector<double> Thickness_vector;
 	data.push_back( Thickness_vector );
 	data.push_back( E_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float E = Emin; E <= Emax; E+=dE )
+	for(double E = Emin; E <= Emax; E+=dE )
 	{
 		try
 		{
@@ -513,7 +513,7 @@ bool get_Thickness_vs_Eout(	StopPow & model ,
 			}
 			else
 			{
-				float Tlast = data[1].back();
+				double Tlast = data[1].back();
 				// do an incremental calculation based on last point:
 				data[0].push_back(E);
 				data[1].push_back( Tlast - model.Thickness(E, fmax(Emin,E-dE)) );
@@ -531,8 +531,8 @@ bool get_Thickness_vs_Eout(	StopPow & model ,
 
 // Get dataset for thickness as a function of energy in given energy out
 bool get_Thickness_vs_Ein(	StopPow & model ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	return get_Thickness_vs_Ein(model,
 								PLOT_DEFAULT_NUM_POINTS,
@@ -543,8 +543,8 @@ bool get_Thickness_vs_Ein(	StopPow & model ,
 // Get dataset for thickness as a function of energy in given energy out
 bool get_Thickness_vs_Ein(	StopPow & model ,
 							int num_points ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	// set energy limits
 	// minimum energy is Eout (thickness -> 0)
@@ -559,25 +559,25 @@ bool get_Thickness_vs_Ein(	StopPow & model ,
 
 // Get dataset for thickness as a function of energy in given energy out
 bool get_Thickness_vs_Ein(	StopPow & model ,
-							float Emin ,
-							float Emax ,
+							double Emin ,
+							double Emax ,
 							int num_points ,
-							float Eout ,
-							std::vector< std::vector<float> > & data )
+							double Eout ,
+							std::vector< std::vector<double> > & data )
 {
 	// clear data
 	data.clear();
 
 	// set up 2 vectors for E and dE/dx
-	std::vector<float> E_vector; std::vector<float> Thickness_vector;
+	std::vector<double> E_vector; std::vector<double> Thickness_vector;
 	data.push_back( Thickness_vector );
 	data.push_back( E_vector );
 
 	// step size:
-	float dE = (Emax-Emin) / ((float)num_points);
+	double dE = (Emax-Emin) / ((double)num_points);
 
 	// now generate the plot values:
-	for(float E = Emin; E <= Emax; E+=dE )
+	for(double E = Emin; E <= Emax; E+=dE )
 	{
 		try
 		{
@@ -589,7 +589,7 @@ bool get_Thickness_vs_Ein(	StopPow & model ,
 			}
 			else
 			{
-				float Tlast = data[1].back();
+				double Tlast = data[1].back();
 				// otherwise do incremental based off of last point
 				data[0].push_back(E);
 				data[1].push_back( Tlast + model.Thickness(E, fmax(E-dE,Emin)) );
